Include bytecode index of method to the experiment results
bci is added to the end of experiment name, making the form <class fqn>:<line number>|<bci>. This helps identify which instruction on the line can be optimized (useful in case there are many "hot" instructions on the same line).
But it worth mentioning that information about bci can be misleading due to imperfect mapping from sampled program counter to bytecode index.