[tools] [progress point finder] Reduce overhead of instrumentation
- use internal structure instead of file. I/O operations on every line visit is too wasteful.
- the same, maybe even furthermore, applies to timestamp. It will be removed with ticks that will increase every new line hit.

To achieve this, an agent loads new class with desired structure. This class is not generated -- the existing class agent is converted to byte array. It can bring soe troubles, but at this time it is very useful -- we do not need to write tons of instrumentation code when it will likely to be changed many times in the future.
Currently agent does not measure frequency, only number of hits. And it seems to be working.